More targeted test harnesses
